roshaaapi/requirements.txt
asgiref==3.3.4
certifi==2021.5.30
chardet==4.0.0
colorama==0.4.4
dj-database-url==0.5.0
Django==3.2.4
django-filter==2.4.0
django-heroku==0.3.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.1
et-xmlfile==1.1.0
gunicorn==20.1.0
httpie==2.4.0
idna==2.10
psycopg2==2.9.1
Pygments==2.9.0
PyJWT==2.1.0
PySocks==1.7.1
python-dateutil==2.8.1
pytz==2021.1
requests==2.25.1
requests-toolbelt==0.9.1
six==1.16.0
sqlparse==0.4.1
urllib3==1.26.6
whitenoise==5.2.0
wincertstore==0.2
